+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/notes/TimelineStoreImpl.kt
@@ -6,34 +6,26 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import net.pantasystem.milktea.api.mastodon.status.TootStatusDTO
-import net.pantasystem.milktea.api.misskey.favorite.Favorite
-import net.pantasystem.milktea.api.misskey.notes.NoteDTO
-import net.pantasystem.milktea.api.misskey.notes.NoteRequest
-import net.pantasystem.milktea.api.misskey.v12.MisskeyAPIV12
 import net.pantasystem.milktea.app_store.notes.InitialLoadQuery
 import net.pantasystem.milktea.app_store.notes.TimelineStore
 import net.pantasystem.milktea.common.PageableState
 import net.pantasystem.milktea.common.StateContent
-import net.pantasystem.milktea.common.paginator.*
+import net.pantasystem.milktea.common.paginator.FuturePagingController
+import net.pantasystem.milktea.common.paginator.PaginationState
+import net.pantasystem.milktea.common.paginator.PreviousPagingController
+import net.pantasystem.milktea.common.paginator.StateLocker
 import net.pantasystem.milktea.common.runCancellableCatching
-import net.pantasystem.milktea.common.throwIfHasError
 import net.pantasystem.milktea.data.api.mastodon.MastodonAPIProvider
 import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.account.page.Pageable
-import net.pantasystem.milktea.model.account.page.SincePaginate
-import net.pantasystem.milktea.model.account.page.UntilPaginate
 import net.pantasystem.milktea.model.instance.MetaRepository
-import net.pantasystem.milktea.model.nodeinfo.NodeInfo
 import net.pantasystem.milktea.model.nodeinfo.NodeInfoRepository
 import net.pantasystem.milktea.model.notes.Note
 import net.pantasystem.milktea.model.notes.NoteDataSource
 import net.pantasystem.milktea.model.notes.NoteRelation
 import net.pantasystem.milktea.model.notes.NoteRelationGetter
-import retrofit2.Response
 import javax.inject.Inject
 
 
@@ -271,313 +263,3 @@ class TimelineStoreImpl(
 sealed interface TimelinePagingBase : PaginationState<Note.Id>, StateLocker
 
 interface StreamingReceivableStore : StateLocker
-
-class TimelinePagingStoreImpl(
-    private val pageableTimeline: Pageable,
-    private val noteAdder: NoteDataSourceAdder,
-    private val getAccount: suspend () -> Account,
-    private val getInitialLoadQuery: () -> InitialLoadQuery?,
-    private val misskeyAPIProvider: MisskeyAPIProvider,
-    private val metaRepository: MetaRepository,
-) : EntityConverter<NoteDTO, Note.Id>, PreviousLoader<NoteDTO>, FutureLoader<NoteDTO>,
-    IdGetter<Note.Id>, TimelinePagingBase, StreamingReceivableStore {
-
-    private val _state =
-        MutableStateFlow<PageableState<List<Note.Id>>>(PageableState.Loading.Init())
-    override val state: Flow<PageableState<List<Note.Id>>>
-        get() = _state
-
-    override val mutex: Mutex = Mutex()
-
-    override suspend fun getSinceId(): Note.Id? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.maxByOrNull {
-            it.noteId
-        }
-    }
-
-    override suspend fun getUntilId(): Note.Id? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.minByOrNull {
-            it.noteId
-        }
-    }
-
-    override suspend fun loadPrevious(): Result<List<NoteDTO>> {
-        return runCancellableCatching {
-            val untilId = getUntilId()?.noteId
-            if (pageableTimeline !is UntilPaginate && untilId != null) {
-                return@runCancellableCatching emptyList()
-            }
-            val builder = NoteRequest.Builder(
-                i = getAccount.invoke().token,
-                pageable = pageableTimeline,
-                limit = LIMIT
-            )
-            val initialLoadQuery = getInitialLoadQuery.invoke()
-            val untilDate = (initialLoadQuery as? InitialLoadQuery.UntilDate)?.date
-            val req = builder.build(
-                NoteRequest.Conditions(
-                    untilId = untilId
-                        ?: (initialLoadQuery as? InitialLoadQuery.UntilId)?.noteId?.noteId,
-                    untilDate = if (untilId == null) untilDate?.toEpochMilliseconds() else null,
-                )
-            )
-            getStore()!!.invoke(req).throwIfHasError().body()!!
-        }
-    }
-
-    override suspend fun loadFuture(): Result<List<NoteDTO>> {
-        return runCancellableCatching {
-            if (pageableTimeline !is SincePaginate) {
-                return@runCancellableCatching emptyList()
-            }
-
-            val builder = NoteRequest.Builder(
-                i = getAccount.invoke().token,
-                pageable = pageableTimeline,
-                limit = LIMIT
-            )
-            val req = builder.build(NoteRequest.Conditions(sinceId = getSinceId()?.noteId))
-            getStore()!!.invoke(req).throwIfHasError().body()!!
-        }
-    }
-
-    override suspend fun convertAll(list: List<NoteDTO>): List<Note.Id> {
-        return list.filter {
-            it.promotionId == null || it.tmpFeaturedId == null
-        }.map {
-            noteAdder.addNoteDtoToDataSource(getAccount.invoke(), it).id
-        }
-    }
-
-    override fun getState(): PageableState<List<Note.Id>> {
-        return _state.value
-    }
-
-    override fun setState(state: PageableState<List<Note.Id>>) {
-        _state.value = state
-    }
-
-    private suspend fun getStore(): (suspend (NoteRequest) -> Response<List<NoteDTO>?>)? {
-        val account = getAccount.invoke()
-        val meta = metaRepository.find(account.normalizedInstanceDomain)
-        val api = misskeyAPIProvider.get(account, meta.getOrNull()?.getVersion())
-        return try {
-            when (pageableTimeline) {
-                is Pageable.GlobalTimeline -> api::globalTimeline
-                is Pageable.LocalTimeline -> api::localTimeline
-                is Pageable.HybridTimeline -> api::hybridTimeline
-                is Pageable.HomeTimeline -> api::homeTimeline
-                is Pageable.Search -> api::searchNote
-                is Pageable.Favorite -> throw IllegalArgumentException("use FavoriteNotePagingStore.kt")
-                is Pageable.UserTimeline -> api::userNotes
-                is Pageable.UserListTimeline -> api::userListTimeline
-                is Pageable.SearchByTag -> api::searchByTag
-                is Pageable.Featured -> api::featured
-                is Pageable.Mention -> api::mentions
-                is Pageable.Antenna -> {
-                    if (api is MisskeyAPIV12) {
-                        (api)::antennasNotes
-                    } else {
-                        throw IllegalArgumentException("antennaはV12以上でなければ使用できません")
-                    }
-                }
-                is Pageable.ChannelTimeline -> {
-                    if (api is MisskeyAPIV12) {
-                        (api)::channelTimeline
-                    } else {
-                        throw IllegalArgumentException("channelはV12以上でなければ使用できません")
-                    }
-                }
-                else -> throw IllegalArgumentException("unknown class:${pageableTimeline.javaClass}")
-            }
-
-        } catch (e: NullPointerException) {
-            null
-        }
-
-    }
-}
-
-
-class FavoriteNoteTimelinePagingStoreImpl(
-    val pageableTimeline: Pageable.Favorite,
-    val noteAdder: NoteDataSourceAdder,
-    val getAccount: suspend () -> Account,
-    val misskeyAPIProvider: MisskeyAPIProvider,
-) : EntityConverter<Favorite, Note.Id>, PreviousLoader<Favorite>, FutureLoader<Favorite>,
-    IdGetter<String>, StateLocker, TimelinePagingBase {
-
-    var favoriteIdAndNoteIdMap = mutableMapOf<Note.Id, String>()
-
-    private val _state =
-        MutableStateFlow<PageableState<List<Note.Id>>>(PageableState.Loading.Init())
-    override val state: Flow<PageableState<List<Note.Id>>>
-        get() = _state
-
-    override val mutex: Mutex = Mutex()
-
-    override suspend fun convertAll(list: List<Favorite>): List<Note.Id> {
-        val account = getAccount.invoke()
-
-        val fabIdAndNoteId = list.map {
-            noteAdder.addNoteDtoToDataSource(account, it.note).id to it.id
-        }
-        favoriteIdAndNoteIdMap.putAll(fabIdAndNoteId.toMap())
-        return fabIdAndNoteId.map {
-            it.first
-        }
-    }
-
-    override suspend fun getSinceId(): String? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.firstOrNull()?.let {
-            favoriteIdAndNoteIdMap[it]
-        }
-    }
-
-    override suspend fun getUntilId(): String? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.lastOrNull()?.let {
-            favoriteIdAndNoteIdMap[it]
-        }
-    }
-
-    override suspend fun loadFuture(): Result<List<Favorite>> {
-        val ac = getAccount.invoke()
-        return runCancellableCatching {
-            misskeyAPIProvider.get(getAccount.invoke()).favorites(
-                NoteRequest.Builder(pageableTimeline, ac.token, limit = LIMIT)
-                    .build(NoteRequest.Conditions(sinceId = getSinceId()))
-            ).throwIfHasError().body()!!
-        }
-    }
-
-    override suspend fun loadPrevious(): Result<List<Favorite>> {
-        return runCancellableCatching {
-            val ac = getAccount.invoke()
-            misskeyAPIProvider.get(getAccount.invoke()).favorites(
-                NoteRequest.Builder(pageableTimeline, ac.token, limit = LIMIT)
-                    .build(NoteRequest.Conditions(untilId = getUntilId()))
-            ).throwIfHasError().body()!!
-        }
-    }
-
-
-    override fun getState(): PageableState<List<Note.Id>> {
-        return _state.value
-    }
-
-    override fun setState(state: PageableState<List<Note.Id>>) {
-        _state.value = state
-    }
-}
-
-
-class MastodonTimelineStorePagingStoreImpl(
-    val pageableTimeline: Pageable.Mastodon,
-    val mastodonAPIProvider: MastodonAPIProvider,
-    val getAccount: suspend () -> Account,
-    val noteAdder: NoteDataSourceAdder,
-    val nodeInfoRepository: NodeInfoRepository,
-) : EntityConverter<TootStatusDTO, Note.Id>, PreviousLoader<TootStatusDTO>,
-    FutureLoader<TootStatusDTO>,
-    IdGetter<Note.Id>, StateLocker, TimelinePagingBase, StreamingReceivableStore {
-
-    private val _state =
-        MutableStateFlow<PageableState<List<Note.Id>>>(PageableState.Loading.Init())
-    override val state: Flow<PageableState<List<Note.Id>>>
-        get() = _state
-    override val mutex: Mutex = Mutex()
-
-    override suspend fun convertAll(list: List<TootStatusDTO>): List<Note.Id> {
-        val account = getAccount()
-        return list.map {
-            noteAdder.addTootStatusDtoIntoDataSource(account, it).id
-        }
-    }
-
-    override suspend fun loadFuture(): Result<List<TootStatusDTO>> = runCancellableCatching {
-        val api = mastodonAPIProvider.get(getAccount())
-
-        when (pageableTimeline) {
-            is Pageable.Mastodon.HashTagTimeline -> api.getHashtagTimeline(
-                pageableTimeline.hashtag,
-                minId = getSinceId()?.noteId
-            )
-            Pageable.Mastodon.HomeTimeline -> api.getHomeTimeline(
-                minId = getSinceId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-            is Pageable.Mastodon.ListTimeline -> api.getListTimeline(
-                minId = getSinceId()?.noteId,
-                listId = pageableTimeline.listId,
-            )
-            is Pageable.Mastodon.LocalTimeline -> api.getPublicTimeline(
-                local = true,
-                minId = getSinceId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-            is Pageable.Mastodon.PublicTimeline -> api.getPublicTimeline(
-                minId = getSinceId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-        }.throwIfHasError().body()!!
-    }
-
-    override suspend fun getSinceId(): Note.Id? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.maxByOrNull {
-            it.noteId
-        }
-    }
-
-    override suspend fun getUntilId(): Note.Id? {
-        return (getState().content as? StateContent.Exist)?.rawContent?.minByOrNull {
-            it.noteId
-        }
-    }
-
-    override fun setState(state: PageableState<List<Note.Id>>) {
-        _state.value = state
-    }
-
-    override fun getState(): PageableState<List<Note.Id>> {
-        return _state.value
-    }
-
-
-    override suspend fun loadPrevious(): Result<List<TootStatusDTO>> = runCancellableCatching {
-        val api = mastodonAPIProvider.get(getAccount())
-        when (pageableTimeline) {
-            is Pageable.Mastodon.HashTagTimeline -> api.getHashtagTimeline(
-                pageableTimeline.hashtag,
-                maxId = getUntilId()?.noteId
-            )
-            Pageable.Mastodon.HomeTimeline -> api.getHomeTimeline(
-                maxId = getUntilId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-            is Pageable.Mastodon.ListTimeline -> api.getListTimeline(
-                maxId = getUntilId()?.noteId,
-                listId = pageableTimeline.listId,
-
-                )
-            is Pageable.Mastodon.LocalTimeline -> api.getPublicTimeline(
-                local = true,
-                maxId = getUntilId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-            is Pageable.Mastodon.PublicTimeline -> api.getPublicTimeline(
-                maxId = getUntilId()?.noteId,
-                visibilities = getVisibilitiesParameter(getAccount())
-            )
-        }.throwIfHasError().body()!!
-    }
-
-    private suspend fun getVisibilitiesParameter(account: Account): List<String>? {
-        val nodeInfo = nodeInfoRepository.find(account.getHost()).getOrNull() ?: return null
-        return if (nodeInfo.type is NodeInfo.SoftwareType.Mastodon.Fedibird) {
-            listOf("public", "unlisted", "private", "limited", "direct", "personal")
-        } else {
-            null
-        }
-    }
-
-}
